change to double

+++ src/main/java/com/example/Config.java
@@ -4,7 +4,7 @@ public class Config {
   private final String connectionString;
   private final String databaseName;
   private final String collectionName;
-  private final int totalDataSizeGB;
+  private final double totalDataSizeGB;
   private final int writePercentage;
   private final int numThreads;
   private final int documentsPerThread;
@@ -13,7 +13,7 @@ public class Config {
       String connectionString,
       String databaseName,
       String collectionName,
-      int totalDataSizeGB,
+      double totalDataSizeGB,
       int writePercentage,
       int numThreads) {
     this.connectionString = connectionString;
@@ -23,7 +23,7 @@ public class Config {
     this.writePercentage = writePercentage;
     this.numThreads = numThreads;
     long totalDocuments =
-        ((long) totalDataSizeGB * 1024 * 1024 * 1024) / MongoDBScalingTest.TARGET_DOCUMENT_SIZE;
+        (long) (totalDataSizeGB * 1024 * 1024 * 1024) / MongoDBScalingTest.TARGET_DOCUMENT_SIZE;
     this.documentsPerThread = (int) (totalDocuments / numThreads);
   }
 
@@ -32,7 +32,7 @@ public class Config {
         System.getenv("MONGODB_URI"),
         System.getenv("MONGODB_DATABASE"),
         System.getenv("MONGODB_COLLECTION"),
-        Integer.parseInt(System.getenv("TOTAL_DATA_SIZE_GB")),
+        Double.parseDouble(System.getenv("TOTAL_DATA_SIZE_GB")),
         Integer.parseInt(System.getenv("WRITE_PERCENTAGE")),
         Integer.parseInt(System.getenv("NUM_THREADS")));
   }
@@ -50,7 +50,7 @@ public class Config {
     return collectionName;
   }
 
-  public int getTotalDataSizeGB() {
+  public double getTotalDataSizeGB() {
     return totalDataSizeGB;
   }
